+++ NEWS.md
@@ -7,6 +7,7 @@ Improvements:
 * Added examples for `autodb`, `dfd`, `gv`, and `rejoin`.
 * Added a reference for removing avoidable attributes for enhanced third normal
 form.
+* Changed `decompose` to return an error if the data.frame doesn't satisfy the functional dependencies implied by the schema. This will return an error when using `decompose` with a schema derived from the same data.frame if any approximate dependencies were included. Previously, using `decompose` or `dfd` with approximate dependencies would result in constructing a database with duplicate key values, since there's currently no handling of approximate dependencies during database construction, and records ignored in approximate dependencies were being kept. This is incorrect behaviour; `decompose` will be added back for approximate dependencies once the package can properly handle them.
 
 Fixes: